feature/community postlist로 컴포넌트 나누기 완료, 드랍박스 구현중
